Calculate an engine score
This is based on Paul Nelson's talk on search relevance. It sources
query and click logs from the TestSearchSatisfaction schema, runs the
queries, and then generates an engine score and a histogram of the
result positions.

Includes a basic integration with scipy.optimize to do a brute force
search over a grid. Graphs the result when optimizing a 1 or 2
dimensional search space.

Also includes an nDCG implementation, but we don't yet have a source
of data to get the scores to use in nDCG with.

Change-Id: I167d29e934b7f048e120f908a28e883a77d61c00