be more explicit with utf-8 handling
When running queries against jawiki the scores were abysmal. This turned
out to be due to problems comparing between the 'ascii' python thought
we read from the sql query providing results, and the 'utf-8' provided
by runSearch.php (by proxy, because it was returning json which encodes
strings as utf-8).

Change-Id: Ib571a084e24ae944a7b733328dbff2f03292b11b