Calculate multiple engine scores
This is still not as generic as would be ideal, it would be nice if we
could calculate different engine scoring metrics (although we don't have
any) at the same time, but for now this allows calculating multiple
metrics such as ndcg@1..20 in one go. We could, for example, then plot
these values for multiple runs and see the differences.

Change-Id: I9748dbbc72e1052ce49c1d77630317216dc01490