Add basic pre-deployment sanity check for MLR
Implements a very simple configuration-driven sanity checker that
ensures some set of urls is in the top 3 results of a given query. The
intention of this script is to build up a small list of queries and
results for each wiki we deploy MLR to and use that list as a smoke
check before pushing a mediawiki-config change to move a new model to
full production usage.

Not sure relforge is the best place for this, or where the configuration
should really go, but I couldn't think of a better place.

Change-Id: Ie29ef99d2e404fe97e3b2e42b17df22b836385d8