Add sanity checks for more languages
I'm not completely sure these are very good. I didn't really try and
figure out what any of the pages were, just chose random pages from the
bottom 2/3 of the top 100 most popular pages (by page views) of the last
month. Then chose sometimes a redirect and sometimes the primary title
as the query string.

We can probably iterate on these if we get any complaints about search
results that users expect at the top but are not.

Change-Id: I39b47b47f03d02907f5d291050465c14ae596287