Update UTF-32 handing and CLI options in analyze_counts
While working on Phab ticket T182708, I needed to improve the handling
in analyze_counts of high-value Unicode characters that aren't
Chinese. There can be emoji in Wiktionary titles—who knew?

It was also convenient to define a non-standard analyzer (i.e., not
"text"), so I added command line options for host, port, index, and
analyzer to analyze_counts.

Note: This isn't specifically part of T182708, just my general
analysis analysis tools, so I'm *not* specifying it as the bug for
this update.

Change-Id: Iff3776b02f50a4f255fc254342d9ee050e00b005