Extract CachedQuery to a separate module for reuse
The autocomplete scoring will require reusing this componenet. Pull
it out of engineScore.py into it's own thing. This component was doing
some debug logging but not through python logging so it was refactored
to use default logging.

Additionally setup pytest and add some very basic tests of
CachedQuery. It's quite painful to test this currently, so tests
are quite limited.

Change-Id: I89dd1ab7d8c4b660568cb6a88f7b1ef1899d42c9