Implement elasticsearch explain parser
First version of explain parser for elasticsearch. Parses explains of
individual hits into structured objects. The feature vectors for all
hits can be extracted and stored somewhere for evaluation. Many hits
(with the same source elasticsearch query) can be merged together
to build up an explain tht contains all parts of the origional query.
This merged query can then be converted into a tensorflow graph,
and the feature vectors extracted earlier can be run through it to
generate new scores for all the hits.

The code is quite messy, but it at least has test cases for most
of the pieces of the wikidata item autocomplete query. It has been
tested against ~500k hits for that query and returns the same
values when running the tensorflow graph as were in the original
explain.

Bug: T208917
Change-Id: If073a239229d5b6f4671598c0be5d9753fb79be1