Add support for multiple rescore queries
This implied a refactoring of how the explain parser generates names.
Previously the query tree and explain were relatively close but
by generalizing te rescore it's now completely different.
Rather than trying to reconstruct meaningful namespace from a BaseExplain
tree we give the query parser a chance to design the name spaces which
is probably closer to what the user sees.
The removal of "tfNorm" is the name path is intentional as I feel it
does not bring much value, but can be easily re-added if it needs to.
The merge of missing rescore queries has not been tested.

Bug: T209485
Change-Id: Ia991861b919b81f916c9bc793cab998eb33a1101