Add jupyter notebooks for wbsearchentities analysis
As part of developing the improved autocomplete profiles for
wbsearchentities a couple notebooks were used and included
here. This is perhaps not the best precedent, there is reusable
code that should be refactored out of these notebooks in the future,
but that can wait until we actually have more notebooks and have
a better grasp on our needs.

Notebooks included:

= wbsearchentities explain tuning analysis

Runs over the reports generated by the tf_autocomplete /
explain_parser functionality of relevance forge. Reports some
information about how the tuning went, what has gotten worse
and what has gotten better

= WikiDataCompletionClicks AB analysis

Operates on the data collected by WikimediaEvents for the
WikiDataCompletionClicks schema in javascript. Calculates
the mean characters typed per lookup and per page load,
the abandonment rate, the mean click position, and the
percentages of clicks@{1,2,3,4}

Bug: T208917
Bug: T211033
Change-Id: Ic12cb2439a98a5367ec8a952b57ad6751c094342