Restructure as multiple python projects
We were mixing plenty of things that were unrelated. Split it all
apart into separate pieces. I'm not certain that everything is
now in a working state, but all the tests that were previously written
still run.

This new structure is a bit verbose, but should allow us to keep many
small projects in this relevance forge repository.

Change-Id: I26d4bc16197d5b2b62dcb7c3dbd72917fbb8c963