Calculate DYM metrics for full text search
Calculates boolean properties of all search sessions in eventlogging
and then presents statistics about the distribution of those
properties across searches. For the moment there are far too many
metrics, as we work with it some more we should focus in on a few
meaningful metrics rather than this host of ~20 somewhat
chosen metrics.

Bug: T228226
Change-Id: Iae0b2119ae76a547d1e6593964eed3523531e294