Pin protobuf to <4
New versions of protobuf are incompatible with the older version
of tensorflow installed here. Pin it to keep using a version prior
to the breaking change (released as 4.21.0, skipping 0-20).

Also includes two minor whitespace changes to make flake8 happy, CI
is pulling a new version since this repo was last updated.

Change-Id: I61da42db068acb256da29d58d4e8e1a2e1b84d03